+++ CHANGELOG.md
@@ -2,10 +2,11 @@
 
 **?.?.? Changes:**
 
-* Added 3 new effects:
+* Added 4 new effects:
   * No sprinting: Disables sprinting for all characters, lasts 30 seconds
   * Everyone is Invisible: Every character on the stage becomes invisible, lasts 30 seconds
   * Revive Dead Characters: Revives all recently killed characters
+  * The Floor is Lava: Every character touching the ground is set on fire
 
 * Duplicate Random Item Stack:
   * Added config option to blacklist specific items from being duplicated